third_party/WebKit/LayoutTests/html/dialog/modal-dialog-scroll-height.html

<!DOCTYPE html>
<style>
body {
  margin: 0;
}
.spacer {
  height: 500px;
}
dialog {
  border: 0;
  margin: 0;
  padding: 1px;
}
</style>
<div class="spacer"></div>
<dialog>
  <div class="spacer"></div>
</dialog>
<script src="../../resources/js-test.js"></script>
<script>

document.querySelector('dialog').showModal();
shouldBe('600', String(document.scrollingElement.scrollHeight));

if (window.testRunner)
  testRunner.dumpAsText();

</script>
